rename custom_property to bpy_data, add internal data folder
