+++ batoms/internal_data/bpy_data.py
@@ -371,58 +371,3 @@ class BatomsObject(bpy.types.PropertyGroup):
     settings_attribute: CollectionProperty(name='settings_attribute',
                                  type=Battribute)
 
-
-
-classes = [
-    Belement,
-    Bspecies,
-    Batom,
-    Battribute,
-    Bcell,
-    Bvolume,
-    Bselect,
-    Bboundary,
-    BatomsCollection,
-    BatomsObject,
-]
-
-
-import importlib
-
-modules = ['bond',
-            'polyhedra',
-            'render',
-            'ribbon']
-
-def enable_module():
-    for key in modules:
-        module = importlib.import_module("batoms.{}".format(key))
-        module.register_class()   
-
-def disable_module():
-    for key in modules:
-        module = importlib.import_module("batoms.{}".format(key))
-        module.unregister_class()   
-        
-from bpy.types import Collection, Object
-
-def register_class():
-    from bpy.utils import register_class
-    for cls in classes:
-        register_class(cls)
-    
-    enable_module()
-    Collection.batoms = PointerProperty(name='Batoms',
-                                        type=BatomsCollection)
-    Object.batoms = PointerProperty(name='Batoms',
-                                    type=BatomsObject)
-    
-
-def unregister_class():
-    from bpy.utils import unregister_class
-    for cls in reversed(classes):
-        unregister_class(cls)
-    disable_module()
-    del Collection.batoms
-    del Object.batoms
-
