+++ batoms/plugins/template/bpy_data.py
@@ -5,7 +5,8 @@ from bpy.props import (StringProperty,
                        CollectionProperty,
                        )
 
-from batoms.custom_property import Base
+from batoms.internal_data import Base
+
 
 
 class TemplateSetting(Base):
