+++ batoms/polyhedra/bpy_data.py
@@ -8,7 +8,8 @@ from bpy.props import (StringProperty,
                        CollectionProperty,
                        )
 
-from batoms.custom_property import Base
+from batoms.internal_data import Base
+
 
 
 class PolyhedraSetting(Base):
