+++ batoms/render/bpy_data.py
@@ -7,7 +7,8 @@ from bpy.props import (StringProperty,
                        CollectionProperty,
                        )
 
-from batoms.custom_property import Base
+from batoms.internal_data import Base
+
 
 class LightSetting(bpy.types.PropertyGroup):
     """
