+++ batoms/ribbon/bpy_data.py
@@ -7,7 +7,7 @@ from bpy.props import (StringProperty,
                        CollectionProperty,
                        )
 
-from batoms.custom_property import Base
+from batoms.internal_data import Base
 
 
 class SheetSetting(Base):
